+++ src/or_fail_functions.php
@@ -12,8 +12,6 @@ declare(strict_types=1);
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  * @since       1.0.6
  */
-use ErrorException as ErrorException;
-use Exception as Exception;
 use Tools\Exception\FileNotExistsException;
 use Tools\Exception\KeyNotExistsException;
 use Tools\Exception\NotDirectoryException;
@@ -28,15 +26,18 @@ if (!function_exists('file_exists_or_fail')) {
      * Checks whether a file or directory exists and throws an exception if the
      *  file does not exist
      * @param string $filename Path to the file or directory
-     * @param string $message The failure message that will be appended to the
+     * @param string|null $message The failure message that will be appended to the
      *  generated message
-     * @param string $exception The exception class you want to set
+     * @param \Throwable|string $exception The exception class you want to set
      * @return void
      * @throws \Tools\Exception\FileNotExistsException
      */
-    function file_exists_or_fail(string $filename, string $message = 'File or directory `%s` does not exist', string $exception = FileNotExistsException::class): void
+    function file_exists_or_fail(string $filename, $message = null, $exception = FileNotExistsException::class): void
     {
-        is_true_or_fail(file_exists($filename), sprintf($message, rtr($filename)), $exception);
+        if (!$message) {
+            $message = sprintf('File or directory `%s` does not exist', $filename);
+        }
+        is_true_or_fail(is_readable($filename), $message, $exception);
     }
 }
 
@@ -48,13 +49,13 @@ if (!function_exists('in_array_or_fail')) {
      * @param array $array The array
      * @param string $message The failure message that will be appended to
      *  the generated message
-     * @param string $exception The exception class you want to set
+     * @param \Throwable|string $exception The exception class you want to set
      * @return void
      * @since 1.2.6
      */
-    function in_array_or_fail($value, array $array, string $message = 'The value is not in array', string $exception = NotInArrayException::class): void
+    function in_array_or_fail($value, array $array, $message = null, $exception = NotInArrayException::class): void
     {
-        if (is_stringable($value) && $message == 'The value is not in array') {
+        if (!$message && is_stringable($value) == 'The value is not in array') {
             $message = sprintf('The value `%s` is not in array', (string)$value);
         }
         is_true_or_fail(in_array($value, $array), $message, $exception);
@@ -66,15 +67,18 @@ if (!function_exists('is_dir_or_fail')) {
      * Tells whether the filename is a directory and throws an exception if the
      *  filename is not a directory
      * @param string $filename Path to the directory
-     * @param string $message The failure message that will be appended to
-     *  the generated message
-     * @param string $exception The exception class you want to set
+     * @param string|null $message The failure message that will be appended to the
+     *  generated message
+     * @param \Throwable|string $exception The exception class you want to set
      * @return void
      * @throws \Tools\Exception\NotDirectoryException
      */
-    function is_dir_or_fail(string $filename, string $message = 'Filename `%s` is not a directory', string $exception = NotDirectoryException::class): void
+    function is_dir_or_fail(string $filename, $message = null, $exception = NotDirectoryException::class): void
     {
-        is_true_or_fail(is_dir($filename), sprintf($message, rtr($filename)), $exception);
+        if (!$message) {
+            $message = sprintf('Filename `%s` is not a directory', $filename);
+        }
+        is_true_or_fail(is_dir($filename), $message, $exception);
     }
 }
 
@@ -84,14 +88,14 @@ if (!function_exists('is_positive_or_fail')) {
      * @param mixed $value The value you want to check
      * @param string $message The failure message that will be appended to the
      *  generated message
-     * @param string $exception The exception class you want to set
+     * @param \Throwable|string $exception The exception class you want to set
      * @return void
      * @since 1.2.5
      * @throws \Tools\Exception\NotPositiveException
      */
-    function is_positive_or_fail($value, string $message = 'The value is not a positive', string $exception = NotPositiveException::class): void
+    function is_positive_or_fail($value, $message = null, $exception = NotPositiveException::class): void
     {
-        if (is_stringable($value) && $message == 'The value is not a positive') {
+        if (!$message && is_stringable($value)) {
             $message = sprintf('The value `%s` is not a positive', (string)$value);
         }
         is_true_or_fail(is_positive($value), $message, $exception);
@@ -103,15 +107,18 @@ if (!function_exists('is_readable_or_fail')) {
      * Tells whether a file exists and is readable and throws an exception if
      *  the file is not readable
      * @param string $filename Path to the file or directory
-     * @param string $message The failure message that will be appended to
+     * @param string|null $message The failure message that will be appended to
      *  the generated message
-     * @param string $exception The exception class you want to set
+     * @param \Throwable|string $exception The exception class you want to set
      * @return void
      * @throws \Tools\Exception\NotReadableException
      */
-    function is_readable_or_fail(string $filename, string $message = 'File or directory `%s` is not readable', string $exception = NotReadableException::class): void
+    function is_readable_or_fail(string $filename, $message = null, $exception = NotReadableException::class): void
     {
-        is_true_or_fail(is_readable($filename), sprintf($message, rtr($filename)), $exception);
+        if (!$message) {
+            $message = sprintf('File or directory `%s` is not readable', $filename);
+        }
+        is_true_or_fail(is_readable($filename), $message, $exception);
     }
 }
 
@@ -124,28 +131,33 @@ if (!function_exists('is_true_or_fail')) {
      * @param mixed $value The value you want to check
      * @param string $message The failure message that will be appended to the
      *  generated message
-     * @param string $exception The exception class you want to set
+     * @param \Throwable|string $exception The exception class you want to set
      * @return void
      * @since 1.1.7
      * @throws \Exception
      */
-    function is_true_or_fail($value, string $message = 'The value is not equal to `true`', string $exception = ErrorException::class): void
+    function is_true_or_fail($value, $message = 'The value is not equal to `true`', $exception = \ErrorException::class): void
     {
         if ($value) {
             return;
         }
 
-        if (func_num_args() < 3 && class_exists($message)) {
-            $exception = new $message();
-        } else {
+        if ($message instanceof \Throwable || (is_string($message) && class_exists($message))) {
+            [$exception, $message] = [$message, 'The value is not equal to `true`'];
+        }
+
+        if (!$exception instanceof \Throwable) {
+            if (!is_string($exception)) {
+                trigger_error('`$exception` parameter must be a string');
+            }
             if (!class_exists($exception)) {
                 trigger_error(sprintf('Class `%s` does not exist', $exception));
             }
             $exception = new $exception($message);
         }
 
-        if (!$exception instanceof Exception) {
-            trigger_error(sprintf('`%s` is not and instance of `Exception`', get_class($exception)));
+        if (!$exception instanceof \Throwable) {
+            trigger_error(sprintf('`%s` is not and instance of `Throwable`', get_class($exception)));
         }
 
         throw $exception;
@@ -157,15 +169,18 @@ if (!function_exists('is_writable_or_fail')) {
      * Tells whether the filename is writable and throws an exception if the
      *  file is not writable
      * @param string $filename Path to the file or directory
-     * @param string $message The failure message that will be appended to
+     * @param string|null $message The failure message that will be appended to
      *  the generated message
-     * @param string $exception The exception class you want to set
+     * @param \Throwable|string $exception The exception class you want to set
      * @return void
      * @throws \Tools\Exception\NotWritableException
      */
-    function is_writable_or_fail(string $filename, string $message = 'File or directory `%s` is not writable', string $exception = NotWritableException::class): void
+    function is_writable_or_fail(string $filename, $message = null, $exception = NotWritableException::class): void
     {
-        is_true_or_fail(is_writable($filename), sprintf($message, rtr($filename)), $exception);
+        if (!$message) {
+            $message = sprintf('File or directory `%s` is not writable', $filename);
+        }
+        is_true_or_fail(is_writable($filename), $message, $exception);
     }
 }
 
@@ -177,16 +192,20 @@ if (!function_exists('key_exists_or_fail')) {
      * If you pass an array of keys, they will all be checked.
      * @param string|int|array $key Key to check or an array of keys
      * @param array $array An array with keys to check
-     * @param string $message The failure message that will be appended to
+     * @param \Throwable|string $message The failure message that will be appended to
      *  the generated message
      * @param string $exception The exception class you want to set
      * @return void
      * @throws \Tools\Exception\KeyNotExistsException
      */
-    function key_exists_or_fail($key, array $array, $message = 'Key `%s` does not exist', $exception = KeyNotExistsException::class)
+    function key_exists_or_fail($key, array $array, $message = null, $exception = KeyNotExistsException::class): void
     {
         foreach ((array)$key as $name) {
-            is_true_or_fail(array_key_exists($name, $array), sprintf($message, $name), $exception);
+            $result = array_key_exists($name, $array);
+            if (!$result && !$message) {
+                $message = sprintf('Key `%s` does not exist', $name);
+            }
+            is_true_or_fail($result, $message, $exception);
         }
     }
 }
@@ -202,16 +221,19 @@ if (!function_exists('property_exists_or_fail')) {
      * @param string $property The name of the property
      * @param string $message The failure message that will be appended to
      *  the generated message
-     * @param string $exception The exception class you want to set
+     * @param \Throwable|string $exception The exception class you want to set
      * @return void
      * @since 1.1.14
      * @throws \Tools\Exception\PropertyNotExistsException
      */
-    function property_exists_or_fail($object, $property, $message = 'Object does not have `%s` property', $exception = PropertyNotExistsException::class)
+    function property_exists_or_fail($object, $property, $message = null, $exception = PropertyNotExistsException::class): void
     {
         foreach ((array)$property as $name) {
+            if (!$message) {
+                $message = sprintf('Object does not have `%s` property', $name);
+            }
             $result = method_exists($object, 'has') ? $object->has($name) : property_exists($object, $name);
-            is_true_or_fail($result, sprintf($message, $name), $exception);
+            is_true_or_fail($result, $message, $exception);
         }
     }
 }